fix cpplint

+++ stan/math/rev/core/var.hpp
@@ -150,7 +150,7 @@ class var {
    *
    * @param x Value of the variable.
    */
-  var(long long x) : vi_(new vari(static_cast<double>(x))) {}  // NOLINT
+  var(int64_t x) : vi_(new vari(static_cast<double>(x))) {}  // NOLINT
 
   /**
    * Construct a variable from the specified arithmetic argument
@@ -159,8 +159,8 @@ class var {
    *
    * @param x Value of the variable.
    */
-  var(unsigned long long x)
-      : vi_(new vari(static_cast<double>(x))) {}  // NOLINT
+  var(uint64_t x) // NOLINT
+      : vi_(new vari(static_cast<double>(x))) {}
 
   /**
    * Construct a variable from the specified arithmetic argument
